Extract shared routing helpers
