+++ packages/apps/browser-app/package.json
@@ -36,6 +36,7 @@
     "@superego/echarts": "workspace:*",
     "@superego/global-types": "workspace:*",
     "@superego/javascript-sandbox-global-utils": "workspace:*",
+    "@superego/routing": "workspace:*",
     "@superego/schema": "workspace:*",
     "@superego/shared-utils": "workspace:*",
     "@tanstack/react-query": "^5.100.14",
