+++ packages/core/routing/src/index.ts
@@ -1,0 +1,8 @@
+export {
+  fromDeepLink,
+  toDeepLink,
+  toHrefFromDeepLink,
+} from "./DeepLinkUtils.js";
+export type { default as Route } from "./Route.js";
+export { CollectionRouteView, RouteName } from "./Route.js";
+export { fromHref, toHref, tryFromHref } from "./RouteUtils.js";
