+++ packages/core/routing/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": ["../../../tsconfig.common.json"],
+  "compilerOptions": {
+    "types": ["node"]
+  }
+}
